Moved classes to proper packages
