+++ src/wildpark/model/animals/RunningAnimal.java
@@ -1,0 +1,5 @@
+package wildpark.model.animals;
+
+public interface RunningAnimal {
+
+}
