+++ src/wildpark/model/animals/WalkingAnimal.java
@@ -1,0 +1,5 @@
+package wildpark.model.animals;
+
+public interface WalkingAnimal {
+
+}
