--- src/wildpark/model/animals/mammals/RunningAnimal.java
@@ -1,5 +1,0 @@
-package wildpark.model.animals.mammals;
-
-public interface RunningAnimal {
-
-}
