--- src/wildpark/model/animals/mammals/WalkingAnimal.java
@@ -1,5 +1,0 @@
-package wildpark.model.animals.mammals;
-
-public interface WalkingAnimal {
-
-}
